move cgcollector includes to own cgcollector namespace to avoid name congestions
