+++ tools/cgcollector2/include/cgcollector2/MetaDataFunctions.h
