+++ tools/cgcollector2/include/cgcollector2/Plugin.h
