+++ tools/cgcollector2/include/cgcollector2/SharedDefs.h
