+++ tools/cgcollector2/include/cgcollector2/helper/Common.h
