+++ tools/cgcollector2/src/CallGraphCollectionAction.cpp
@@ -4,14 +4,14 @@
  * https://github.com/tudasc/metacg/LICENSE.txt
  */
 
-#include "CallGraphCollectionAction.h"
-#include "CallGraphNodeGenerator.h"
-#include "Plugin.h"
-#include "SharedDefs.h"
 #include "metacg/MCGBaseInfo.h"
-#include "metadata/Internal/ASTNodeMetadata.h"
-#include "metadata/Internal/AllAliasMetadata.h"
-#include "metadata/Internal/FunctionSignatureMetadata.h"
+#include "cgcollector2/CallGraphCollectionAction.h"
+#include "cgcollector2/CallGraphNodeGenerator.h"
+#include "cgcollector2/Plugin.h"
+#include "cgcollector2/SharedDefs.h"
+#include "cgcollector2/metadata/Internal/ASTNodeMetadata.h"
+#include "cgcollector2/metadata/Internal/AllAliasMetadata.h"
+#include "cgcollector2/metadata/Internal/FunctionSignatureMetadata.h"
 
 #include "metacg/MCGManager.h"
 #include "metacg/io/VersionFourMCGWriter.h"
